jobs/kpi-forecasting/requirements.txt
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
BigQuery-Python==1.15.0
black==20.8b1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cmdstanpy==1.0.7
convertdate==2.4.0
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.28
ephem==4.1.3
flake8==3.8.4
google-api-core==1.31.5
google-api-python-client==2.38.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
google-auth==1.35.0
google-cloud-bigquery-storage==1.0.0
google-cloud-bigquery==1.27.2
google-cloud-core==1.7.2
google-cloud-storage==1.31.0
google-crc32c==1.3.0
google-resumable-media==1.3.3
google==3.0.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
hijri-converter==2.2.3
holidays==0.16
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
Jinja2==2.11.2
joblib==1.2.0
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
more-itertools==8.6.0
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.2.0
packaging==20.7
packaging==21.3
pandas-gbq==0.13.2
pandas==1.3.5
paramiko==2.9.2
pathspec==0.8.1
Pillow==9.0.1
plotly==4.9.0
pluggy==0.13.1
protobuf==3.19.4
py==1.10.0
pyarrow==7.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.21
pydata-google-auth==1.3.0
pyflakes==2.2.0
PyMeeus==0.5.11
PyNaCl==1.5.0
pyparsing==2.4.7
pytest-black==0.3.11
pytest-flake8==1.0.6
pytest==6.0.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2020.11.13
requests-oauthlib==1.3.1
requests==2.27.1
retrying==1.3.3
rsa==4.8
scikitlearn==1.1.2
setuptools-git==1.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
statsforecast=0.7.1
statsmodels==0.13.2
storage==0.0.4.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.63.0
typed-ast==1.5.4
typing-extensions==3.7.4.3
ujson==5.1.0
uritemplate==4.1.1
urllib3==1.26.8